% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_organs.R
\name{count_organs}
\alias{count_organs}
\title{Title Counting Plant Organs}
\usage{
count_organs(x)
}
\arguments{
\item{x}{This is a data.frame or tibble of plant uses.
needs to include columns "organ.rough".}
}
\description{
THis function counts the numer of plant organs used.
}
\examples{

data("plant_use")

count_organs(x = plant_use)

}
