% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantuse-package.R
\docType{data}
\name{plant_use}
\alias{plant_use}
\title{plant_use}
\format{
A data with 7 variables:
\describe{
\item{accepted.name}{character. The accepted taxonomic name of the species}
\item{family}{the plant family name}
}
}
\usage{
plant_use
}
\description{
A dataset of plant uses.
}
\examples{

data("plant_use")

}
\keyword{datasets}
