% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_summary.R
\name{report_summary}
\alias{report_summary}
\title{Report some summary stats}
\usage{
report_summary(data, measure, na.rm = TRUE, quantiles = NULL)
}
\arguments{
\item{data}{A dataset}

\item{measure}{The measure you want reported}

\item{na.rm}{Whether or not to remove NAs}

\item{quantiles}{The quantiles that should be included in the output}
}
\description{
Report some summary stats
}
